added buggy texture support for one side

+++ srcs/math_helpers.c
@@ -6,24 +6,24 @@
 /*   By: fheld <fheld@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/04 19:45:52 by fheld             #+#    #+#             */
-/*   Updated: 2023/10/07 14:49:52 by fheld            ###   ########.fr       */
+/*   Updated: 2023/10/08 13:31:42 by fheld            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
 
 /**
- * converts the integer points to floats and
- * calculates the distance from a to b in floats
+ * converts the integer points to doubles and
+ * calculates the distance from a to b in doubles
  * @param a fist point
  * @param b second point
 */
-float	dist(t_int_p2 a, t_int_p2 b)
+double	dist(t_int_p2 a, t_int_p2 b)
 {
-	float	ax;
-	float	ay;
-	float	bx;
-	float	by;
+	double	ax;
+	double	ay;
+	double	bx;
+	double	by;
 
 	ax = a.x;
 	ay = a.y;
